refactor: unify certificates component
